+++ src/evm/mod.rs
@@ -25,6 +25,7 @@ pub use coverage::{
 };
 pub use forkdb::ForkDBConfig;
 pub use forkdb::MockTransport;
+pub use forkdb::RpcStats;
 pub use result::TransactionResult;
 pub use trace::{
     CallFrame, CallFrameKind, Evmole, StorageChange, StorageChangeInfo, StorageType, Trace,